+++ packages/kit/src/types/global-private.d.ts
@@ -11,6 +11,8 @@ declare global {
 	const __SVELTEKIT_SERVER_TRACING_ENABLED__: boolean;
 	/** true if corresponding config option is set to true */
 	const __SVELTEKIT_EXPERIMENTAL__REMOTE_FUNCTIONS__: boolean;
+	/** True if `config.kit.experimental.forkPreloads` is `true` */
+	const __SVELTEKIT_FORK_PRELOADS__: boolean;
 	/** True if `config.kit.router.resolution === 'client'` */
 	const __SVELTEKIT_CLIENT_ROUTING__: boolean;
 	/** True if `config.kit.router.type === 'hash'` */